Fixed allocation bug
Fixed an allocation that was occurring in add_data. This seems to be due to differences in LinRange between Julia versions 1.5 and 1.8. Namely, version 1.8 takes two type parameters, meaning that the code written in Julia 1.5 could not be type inferred. I have added some parametric types to allow for both posibilities